Add empty class definition to enumeration
